Consolidate 'appliance', 'utensil' and 'vessel' recipe direction models into a single 'equipment' model
