+++ reciperadar/models/recipes/direction.py
@@ -2,9 +2,7 @@ from xml.etree import ElementTree
 
 from reciperadar import db
 from reciperadar.models.base import Storable
-from reciperadar.models.recipes.appliance import DirectionAppliance
-from reciperadar.models.recipes.utensil import DirectionUtensil
-from reciperadar.models.recipes.vessel import DirectionVessel
+from reciperadar.models.recipes.equipment import DirectionEquipment
 
 
 class RecipeDirection(Storable):
@@ -17,52 +15,10 @@ class RecipeDirection(Storable):
     index = db.Column(db.Integer)
     description = db.Column(db.String)
     markup = db.Column(db.String)
-    appliances = db.relationship(
-        'DirectionAppliance',
+    equipment = db.relationship(
+        'DirectionEquipment',
         passive_deletes='all'
     )
-    utensils = db.relationship(
-        'DirectionUtensil',
-        passive_deletes='all'
-    )
-    vessels = db.relationship(
-        'DirectionVessel',
-        passive_deletes='all'
-    )
-
-    @staticmethod
-    def _build_item(item, category, category_class):
-        item_classes = set(item.attrib.get('class', '').split())
-        if 'equipment' not in item_classes:
-            return
-        if category not in item_classes:
-            return
-        doc = {category: item.text}
-        return category_class.from_doc(doc)
-
-    @staticmethod
-    def _parse_equipment(markup):
-        equipment = {
-            'appliances': [],
-            'utensils': [],
-            'vessels': [],
-        }
-        if not markup:
-            return equipment
-
-        category_classes = {
-            'appliances': DirectionAppliance,
-            'utensils': DirectionUtensil,
-            'vessels': DirectionVessel,
-        }
-
-        doc = ElementTree.fromstring(f'<xml>{markup}</xml>')
-        for item in doc.findall('mark'):
-            for category in equipment:
-                cls = category_classes[category]
-                obj = RecipeDirection._build_item(item, category[:-1], cls)
-                equipment[category].append(obj) if obj else None
-        return equipment
 
     @staticmethod
     def from_doc(doc, matches=None):
@@ -73,19 +29,16 @@ class RecipeDirection(Storable):
             index=doc['index'],
             description=doc['description'],
             markup=doc['markup'],
-            **equipment
+            equipment=[
+                DirectionEquipment.from_doc(entity)
+                for entity in doc['entities']
+                if entity.get('type') == 'equipment'
+            ],
         )
 
     def to_doc(self):
         data = super().to_doc()
         data['equipment'] = [
-            {'equipment': appliance.appliance}
-            for appliance in self.appliances
-        ] + [
-            {'equipment': utensil.utensil}
-            for utensil in self.utensils
-        ] + [
-            {'equipment': vessel.vessel}
-            for vessel in self.vessels
+            equipment.to_doc() for equipment in self.equipment
         ]
         return data
