+++ reciperadar/models/recipes/equipment.py
@@ -1,0 +1,22 @@
+from reciperadar import db
+from reciperadar.models.base import Storable
+
+
+class DirectionEquipment(Storable):
+    __tablename__ = 'direction_equipment'
+
+    fk = db.ForeignKey('recipe_directions.id', ondelete='cascade')
+    direction_id = db.Column(db.String, fk, index=True)
+
+    id = db.Column(db.String, primary_key=True)
+    name = db.Column(db.String)
+    category = db.Column(db.String)
+
+    @staticmethod
+    def from_doc(doc):
+        equipment_id = doc.get('id') or DirectionEquipment.generate_id()
+        return DirectionEquipment(
+            id=equipment_id,
+            name=doc['name'],
+            category=doc['category'],
+        )
